Closes #320: Ensure large files are used as metadata cache input
The actual metadata cache directory is usually (except on initialization)
created with Pig union, using a common union workflow.
As union is a map-only operation we cannot control the output file sizes
by changing the number of reducers, but we can use the pig.maxCombinedSplitSize
property to make Pig mappers combine the inputs.

Add a parameter for setting the property to the union workflow (defaulting to standard
Pig behaviour).
Use it in cache workflows to make union create large files, up to 320MB.