Closes #522: Affiliation matching is failing due to running beyond physical memory limits
Increase value of `spark.yarn.driver.memoryOverhead` so that more memory
is allocated from YARN for the driver and the workflow finishes.

The direct cause of the failure must be related to non-heap memory use since
Spark ran the driver with -Xmx1024 and it still used more than its 1.5GB
allocated with default overhead value.