Add Jenkinsfile.dockerized for running on the new Jenkins instance
Based on the current Jenkinsfile but uses docker agents with a custom
image providing Maven 3/JDK 8 and Python 2/3 with APSW.

The image also contains /etc/passwd file with an entry for the UID
that it is run as by Jenkins (1000).
Lack of this entry actually would cause a few tests to fail (because of
getpwuid being used by JVM in some place).

Set user.home to the workspace directory (the one in host passwd is not suitable).

Add pollSCM trigger until github-webhook is handled by the new
Jenkins.

Use direct credentials support for the sonar token and hardcode the sonar
URL (actually the same as the sonar name hardcoded in the original
Jenkinsfile), because the sonar plugin is unavailable.

Split build into Build and Sonar stages (the latter only for the
master branch) instead of handling the condition in the shell script.
Gives a nicer display but requires adding -DskipTests -DskipITs to the
second maven run.

In the build step do install, not package - apparently the original
Jenkinsfile actually depends on the installation being the sideeffect
of another build (without install iis-wf-primary tries to download
eu.dnetlib.iis:iis-wf-report:jar:tests artifact from artifactory, but
nothing is actually published there).

Replace the unavailable warnings step with recordIssues from the
warnings-ng plugin.