Closes #1327: Reduce the number of false positives in affiliation matching outcome
Adjusting CommonAffSectionWordsVoter configuration by:
* setting minNumberOfWordsInAffSection = 2, which requires at least two words in a single affiliation section to be a match candidate
* excluding matches with different country codes, to be applied only country code when set in both organization and affiliation

Alighing unit tests and voter trust level. Supplementing test suite with more cases.

We are requiring at least two words in aff section because Jaro-Winkler similarity level is set to `0.85` and having single common word enough to a match considered as valid is quite risky and introduced quite a lot of false positives which at first glance were difficult to explain (e.g. "IN" and "INRA" gives `0.866` similarity and finding "in" word in organization name was enough to vote for a match). Increasing similarity level threshold eliminated too many valid matches and I decided to leave it as it is.

Excluding matches with different country codes was a rather safe call and eliminated most of the false positives reported in https://support.openaire.eu/issues/7392 (8 out of 10, one was not reproducible). It is important this rule is not applied when country code is unknown on either side: organization or affiliation.

This turned out to be quite optimal solution which proved to remove only `33k` of relations (out of `12M`) and when analyzing random sample of 100 removed elements over 80% of them were false positives. The properly matched cases were a poor quality affiliation or organization matched by a sheer luck of a single-word aff section having a common word (e.g. university or an organization abbreviation).

Some of the expected matches were not matched because of the country mismatch due to an invalid country defined on organization side but there is expected another organization having the country code set which should be properly matched, e.g. "University of California" organizations having invalid country code set has ~500 other instances of organization with similar name and a valid country code it is expected to be successfully matched with. No need to match with a poorly defined organization instance.

This change is relevant to `DocOrgRelationMatcher` only (the only matcher unitizing `CommonAffSectionWordsVoter`) so it is bound to doc->proj and proj->org relations based clusterization only.