Do not run SlowTest tests during regular Jenkins builds
As discussed in issue #1329 we do not want to run tests marked as
slow, some of which take dozens of minutes to finish, during regular
IIS builds. It is enough that they are run together with integration
tests and during deployment builds.

So add -DotherExcludedGroups=SlowTest to the maven invocation.

Additional not strictly related small fix - skipTests implies skipITs
so do not define both in the maven invocation for sonar.