workaround for bug with UDF strsplit
strsplit is based on an external lib (python dsv) which was integrated with madis. In this commit, we replace strsplit function call with an equivalent workaround. Debugging of strsplit is more complex since the bug is in third parties code (a csv reader which reads delimited input, this should be replaced with a python3 library).