Closes #1417: Introduce missing integration test for avro to protobuf workflow
This commit introduces missing test verifying correctness of avro -> protobuf transformation workflow used in documentssimilarity workflow as an "interface" between IIS and Coansys workflows.

The major difficulty was to implement a testing consumer verifying whether produced protobuf outcome is built in line with the predefined expectations.
Usually we rely on a simple java class (Process in the IIS nomenclature) but it turned out to be rather cumbersome to veriify protocol buffer record this way. A lot of new protobuf-related code turned out to be required to be written in order to introduce protobuf-compatible TestingConsumer.
This is why less generic, spark based but very simple PersonProtoConsumer was implemented which does the testing consumer job.