Closes #1426: Run IIS experiments by relying on spark 3.4 version
WIP.

Fixing task serialization issue by upgrading avro dependency from 1.8.10 to 1.11.1 which is already a part of sharelib342. This induced the requirement to align JsonConverter with the new code and one of the requirements to move it to a different package due to limited visibility of one of the crucial methods.

Further logging system dependency alignment to make unit tests output produced on console visible.