Closes #1469: Reimplement software mining algorithm to rely on spark while joining mining output with SoftwareHeritage links
Introducing a dedicated spark job, SoftwareHeritageUrlsMatchingJob, responsible for performing joining operation between the mining output and the links coming from SoftwareHeritage.
It is meant to be injected in the oozie workflow after the softwarepost.sql Madis script execution and take over the joining operation which was removed from the madis script in the prior commit.

SoftwareHeritageUrlsMatchingJob was covered with unit and integration tests.