Closes #1466: Fix the JATS XML parsing for affiliations coming from Springer
Fixing unit test revealed lack of proper support for a variant of a semi-complex affiliation (no content-type explicitly defined) having multiple institution elements with separator defined outside of those elements.

Solution might not seem that simple because of the wasInstitutionTagAlreadyClosedForAff flag which is required because concatenated institution text might either come from another institution element but it can also be a continuation of an already opened element (this could be the case for the very long institution names).