Closes #1466: Fix the JATS XML parsing for affiliations coming from Springer
Improving wasInstitutionTagAlreadyClosedForAff flag handling: we need to set it to false whenever consuming a first batch of characters because it might be a long institution text and the 2nd and subsequent execution of characters() method call for a given element should simply concatenate text without any separators.
This could happen when we processed the first institution tag already and then started to process another (potentially very long) institution when multiple characters() method call could be executed.