Closes #1469: Reimplement software mining algorithm to rely on spark while joining mining output with SoftwareHeritage links
Introducing a dedicated spark job, SoftwareHeritageUrlsMatchingJob, responsible for performing joining operation between the mining output and the links coming from SoftwareHeritage.
This new spark job was injected in the oozie workflow after the modified softwarepost.sql Madis script execution and took over the joining operation which was removed from the madis script.

Removing sqlite_builder workflow along with the classes responsible for building SQLite with SoftwareHeritage origins. Removing this phase from the softwareurl_cache_builder workflow poducing avro datastore with SoftwareHeritageOrigin records from now on. Relying on SoftwareHeritageOrigin avro datastore in the main software mining workflow and providing it to the newly introduced script rewritten in spark.

Renaming TaraReferenceExtractionIOUtils to ReferenceExtractionIOUtils and moving this class to a common package.