Closes #1516: Make PMC ingestion compliant with JATS records having article element defined with a namespace
Text extraction converter, which is a submodule of the JATS parser, was modified in a way it accepts any namespace for article element including an empty namespace. This is now in line with the PMC metadata extraction module which is namespace agnostic. Both cases were proved by the newly added unit tests.