Closes #1533: Introduce caching for the fuzzy citation matching algorithm
Introducing input filtering in CitationMatchingInputTransformerJob by relying on cache holding Citations records. Whenever an entry for a given id is found all references for that publications are removed so they are not involved in matching while the main entities remain so they could still be matching candidates (to have other references matched with those entities).

Supplementing CitationMatchingInputTransformerJobTest tests suite with the case involving cache.