Closes #1554: Implement a workflow responsible for importing HTML landing pages obtained by the PDF Aggregation System and sending them to the mining algorithm
Implementing `HtmlLandingPagesExtractionJob` spark job responsible for parsing the input CSV file with HTML landing pages metadata, accessing S3 and extracting tar.gz packages (only the ones listed in the mapping file) and reading the HTML files content.
Input CSV file is expected to have a header defined and contain the following relevant set of fields:
* id: the id of this html-record in the pdfaggregation-DB
* pid_type: the type of the PID (doi, handle, etc.). It may be NULL.
* pid: the PID value. It may be NULL. If it's found then the DOI is preferred over other types of PIDs
* archive_s3_location: the S3 location of the archive which contains multiple html-files inside
* html_filename: the name of the html file in that record
* html_size: the size of the html file
It is optimized in a way records are grouped by archive_s3_location field before files are extracted from each archive. Job accepts max_html_page_size threshold to filter out the large files which are not meant to be parsed.
The job produces `DocumentTextWithDOI` avro records with `id`, `doi`, `text` triples.

Implementing `AffroJob` which is a simple wrapper for the `run_affro.sh` shell script which is meant to be run on datanodes and which accepts `DocumentTextWithDOI` records encoded as JSON at input. The conversion between avro and JSON format is done by the `AffroJob` spark wrapper.
`run_affro.sh` script currently just propagates at output whatever was sent at input (just to prove the wrapper code is working as expected) and it is meant to be filled with an appropriate affro code.
Boths jobs produce reports with basic metrics such as the number of imported and exported records (it may differ due to the html page size threshold).

Introducing oozie workflows for each of the modules mentioned above along with `affro/main` uber-workflow binding two workflows into one chain.