Closes #1529: Prepare a metadata extraction cache remover workflow responsible for a pre-selected cache entries removal
Introducing `GenericExistenceFilterJob` responsible for matching `Identifier` datastore with any arbitrary datastore and producing two datastores with matched and unmatched records.
Spark job is supplemented with a dedicated integration test case.

Cache removal logic is defined in the cache remover workflow and relies on the `GenericExistenceFilterJob` and other shared components which were already used in cache building process.

`AvroSchemaGenerator` class was extended with two utility methods simplifying the process of obtaining schema for a given avro class.