Closes #1577: Avoid placing temporary errors related to communication with Grobid as permanent faults in the cache
Introducing dedicated transient fault datastore in order to preserve transient errors for further analysis. It is impossible to do this efficiently by relying on logs alone.

Handling `SocketTimeoutException` thrown when communicating with the Grobid server as transient error.

Increasing default read and connection timeout value from 1 minute to 10 minutes in order to avoid excessive errors when waiting for Grobid results.

Fixing the bug of not writing processing time for a given publication whenever error was handled.